Fixed refcounting for function parameters
Bug: 25233132

For each function parameter that is of a rs object type, introduces
a local variable and sets its value to the paramemter.
This way, the sys ref count of the object in the parameter is
incremented, preventing the object being deleted if the parameter
is overwritten inside the function.

rsClearObj() is called on the newly introduced variable at the end
of the function.

Test: slang lit tests and RSTest

Change-Id: Ic271c5e279b1f398ff233771b28d0acc5ac47c34